requirements.txt
flet==0.8.1
openai-whisper==20230314
